chore: isolate fingerprint
