+++ static/src/fingerprinting.ts
@@ -1,31 +1,13 @@
-import * as FingerprintJS from "@fingerprintjs/fingerprintjs";
 import { ethers } from "ethers";
 import { keccak256 } from "js-sha3";
-
-// Function to get the device fingerprint
-async function getDeviceFingerprint(): Promise<string> {
-  try {
-    // Initialize an agent at application startup.
-    const fpPromise = FingerprintJS.load();
-
-    // Get the visitor identifier when you need it.
-    const fp = await fpPromise;
-    const result = await fp.get();
-
-    // The visitor identifier:
-    return result.visitorId;
-  } catch (error) {
-    console.error("Error getting device fingerprint:", error);
-    throw new Error("Failed to get device fingerprint.");
-  }
-}
+import { getDeviceFingerprint } from "./ethereum/key-gen/fingerprint";
 
 // Function to derive a deterministic Ethereum private key
 function deriveEthereumPrivateKey(fingerprint: string, userId: string, challenge: string): string {
   const dataToHash = fingerprint + userId + challenge;
   const hash = keccak256(dataToHash);
   // Ensure the hash is 64 characters long for hexZeroPad to pad it correctly
-  return ethers.utils.hexZeroPad(`0x${hash}`, 32);
+  return ethers.zeroPadBytes(`0x${hash}`, 32);
 }
 
 // Main function to get the fingerprint and derive the private key
